For rx admin

+++ src/magic/delete/magic.delete.ts
@@ -6,6 +6,8 @@ import { DeleteDirectiveService } from 'src/directive/delete-directive.service';
 import { MagicService } from 'src/magic-meta/magic.service';
 import { SchemaService } from 'src/schema/schema.service';
 import { AbilityType } from 'src/entity-interface/AbilityType';
+import { RxEventGateway } from 'src/rx-event/rx-event.gateway';
+import { RxEventType } from 'src/rx-event/rx-event';
 
 export class MagicDelete {
   constructor(
@@ -14,6 +16,7 @@ export class MagicDelete {
     private readonly deleteDirectiveService: DeleteDirectiveService,
     public readonly schemaService: SchemaService,
     private readonly magicService: MagicService,
+    protected readonly rxEventGateway: RxEventGateway,
   ) {}
 
   async delete(json: any) {
@@ -30,8 +33,21 @@ export class MagicDelete {
     ).parse(json);
     for (const meta of deleteMetas) {
       deletedInstances[meta.entity] = await this.deleteOne(meta);
+      await this.emitEvent(meta.entity, deletedInstances[meta.entity]);
     }
     return deletedInstances;
+  }
+
+  private async emitEvent(entity: string, ids: number[]) {
+    const entityMeta = this.schemaService.getEntityMetaOrFailed(entity);
+    if (entityMeta.eventable) {
+      this.rxEventGateway.broadcastEvent({
+        eventType: RxEventType.InstanceDeleted,
+        entity: entity,
+        ownerId: this.magicService.me?.id,
+        ids: ids,
+      });
+    }
   }
 
   private getCombinationRelationNames(entityName: string) {
@@ -119,7 +135,7 @@ export class MagicDelete {
       meta.entity,
     );
 
-    if (relationMetas && relationMetas.length > 0) {
+    if (relationMetas && relationMetas.length > 0 && !meta.isSoft) {
       for (const instance of instances) {
         for (const relationMeta of relationMetas) {
           //解除所有关联关系，防止外键约束
@@ -129,13 +145,16 @@ export class MagicDelete {
       }
     }
 
-    if (combinationInstances) {
+    if (combinationInstances && !meta.isSoft) {
       await this.magicService.delete(combinationInstances);
     }
 
-    meta.ids &&
-      meta.ids.length > 0 &&
-      (await entityRepository.delete(meta.ids));
+    if (meta.ids && meta.ids.length > 0) {
+      meta.isSoft
+        ? await entityRepository.softDelete(meta.ids)
+        : await entityRepository.delete(meta.ids);
+    }
+
     return instances.map((entity: any) => entity.id);
   }
